+++ StageRole.cs
@@ -42,6 +42,7 @@ internal enum StageRole
     Imitator,
     Avenger,
     Brawler,
+    Influenza,
     Superbot = 1001,
     Disaster = 1002
 }